Revise build steps again
After inspecting the makefiles again, I realized that the default goal
is not `all`... it is actually `lib`. This fact explains why running
`make` without a target did not result in a successful build.

Running `make all` for both 32-bit and 64-bit builds is the appropriate
thing to do, so make that change here.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>